update to vendor version 5.4.6

+++ src/common/tuklib_mbstr_fw.c
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-/// \file       tuklib_mstr_fw.c
+/// \file       tuklib_mbstr_fw.c
 /// \brief      Get the field width for printf() e.g. to align table columns
 //
 //  Author:     Lasse Collin
@@ -27,5 +27,5 @@ tuklib_mbstr_fw(const char *str, int columns_min)
 	if (width < (size_t)columns_min)
 		len += (size_t)columns_min - width;
 
-	return len;
+	return (int)len;
 }
